Simplify store initialisation.
