+++ .env
@@ -1,8 +1,10 @@
+# React app settings. Restart the process for changes to take effect.
+
 # Set to connect to a Coniql server.
 # REACT_APP_CONIQL_SOCKET=localhost:1000
 
 REACT_APP_BASE_URL=http://localhost:3000
 REACT_APP_SIMULATION_TIME=100
-REACT_APP_PROFILER_ENABLED=true
+REACT_APP_PROFILER_ENABLED=false
 REACT_APP_THROTTLE_PERIOD=100
-REACT_APP_LOG_LEVEL=debug
+REACT_APP_LOG_LEVEL=info
