+++ src/app.tsx
@@ -2,13 +2,9 @@ import React, { Profiler } from "react";
 import "./app.css";
 import { Provider } from "react-redux";
 import { BrowserRouter, Redirect, Switch } from "react-router-dom";
-import { getStore, initialiseStore } from "./redux/store";
+import { store } from "./redux/store";
 import log, { LogLevelDesc } from "loglevel";
 import { lightTheme, darkTheme, ThemeContext } from "./themeContext";
-import { SimulatorPlugin } from "./connection/sim";
-import { ConiqlPlugin } from "./connection/coniql";
-import { ConnectionForwarder } from "./connection/forwarder";
-import { Connection } from "./connection/plugin";
 import { EmbeddedDisplay } from "./ui/widgets";
 import { BaseUrlContext } from "./baseUrl";
 import { onRenderCallback } from "./profilerCallback";
@@ -17,13 +13,6 @@ import { Header } from "./ui/components/Header/header";
 import { Footer } from "./ui/components/Footer/footer";
 
 const baseUrl = process.env.REACT_APP_BASE_URL ?? "http://localhost:3000";
-const SIMULATION_TIME = parseFloat(
-  process.env.REACT_APP_SIMULATION_TIME ?? "100"
-);
-const THROTTLE_PERIOD = parseFloat(
-  process.env.REACT_APP_THROTTLE_PERIOD ?? "100"
-);
-const CONIQL_SOCKET = process.env.REACT_APP_CONIQL_SOCKET;
 
 log.setLevel((process.env.REACT_APP_LOG_LEVEL as LogLevelDesc) ?? "info");
 
@@ -35,22 +24,6 @@ function applyTheme(theme: any): void {
 }
 
 const App: React.FC = (): JSX.Element => {
-  const simulator = new SimulatorPlugin(SIMULATION_TIME);
-  const fallbackPlugin = simulator;
-  const plugins: [string, Connection][] = [
-    ["sim://", simulator],
-    ["loc://", simulator],
-    ["", fallbackPlugin]
-  ];
-  if (CONIQL_SOCKET !== undefined) {
-    const coniql = new ConiqlPlugin(CONIQL_SOCKET);
-    plugins.unshift(["pva://", coniql]);
-    plugins.unshift(["ca://", coniql]);
-    plugins.unshift(["ssim://", coniql]);
-  }
-  const plugin = new ConnectionForwarder(plugins);
-  initialiseStore(plugin, THROTTLE_PERIOD);
-  const store = getStore();
   const { dark } = React.useContext(ThemeContext);
   applyTheme(dark ? darkTheme : lightTheme);
